Move ThemeProvide to app.js to apply it everywhere
