+++ front/src/App.js
@@ -7,6 +7,10 @@ import axios from 'axios';
 // Custom components
 import { HashRouter as Router, Switch, Route } from 'react-router-dom';
 import { CookiesProvider, withCookies } from 'react-cookie';
+import {
+  createMuiTheme,
+  ThemeProvider,
+} from '@material-ui/core/styles';
 import SignIn from './signin/SignIn';
 import ResetPassword from './resetPW/ResetPassword';
 import RenewPassword from './renewPW/RenewPassword';
@@ -22,11 +26,20 @@ import TrackCRUD from './edittracks/tracks';
 import Monthview from './monthview/Monthview';
 import Statistics from './statistics/Statistics';
 import EmailSettings from './EmailSettings/EmailSettings';
-import { Raffle } from "./raffle/raffle";
+import { Raffle } from './raffle/raffle';
 
 // React router. Hashrouter, because normal router won't work in apache
 
 import { validateLogin } from './utils/Utils';
+
+// TO DO: Move this to scss if possible.
+const theme = createMuiTheme({
+  palette: {
+    primary: {
+      main: '#555555',
+    },
+  },
+});
 
 /*
    The main component of the whole project.
@@ -62,35 +75,37 @@ class App extends Component {
     return (
       <CookiesProvider>
         <Router>
-          <div className="App">
-            <header className="App-header">
-              <Nav />
-              <Switch>
-                <Route exact path="/" component={Weekview} />
-                <Route exact path="/signin" component={SignIn} />
-                <Route
-                  path="/signin/reset-password"
-                  component={ResetPassword}
-                />
-                <Route
-                  path="/renew-password/:token?"
-                  component={RenewPassword}
-                />
-                <Route path="/dayview/:date?" component={Dayview} />
-                <Route path="/weekview" component={Weekview} />
-                <Route path="/monthview" component={Monthview} />
-                <Route path="/trackview/:date?/:track?" component={Trackview} />
-                <Route path="/scheduling/:date?" component={Scheduling} />
-                <Route path="/tablet" component={RangeOfficerView} />
-                <Route path="/profile" component={Profile} />
-                <Route path="/usermanagement" component={UserManagementView} />
-                <Route path="/supervisor-raffle" component={Raffle} />
-                <Route path="/tracks" component={TrackCRUD} />
-                <Route path="/email-settings" component={EmailSettings} />
-                <Route path="/statistics" component={Statistics} />
-              </Switch>
-            </header>
-          </div>
+          <ThemeProvider theme={theme}>
+            <div className="App">
+              <header className="App-header">
+                <Nav />
+                <Switch>
+                  <Route exact path="/" component={Weekview} />
+                  <Route exact path="/signin" component={SignIn} />
+                  <Route
+                    path="/signin/reset-password"
+                    component={ResetPassword}
+                  />
+                  <Route
+                    path="/renew-password/:token?"
+                    component={RenewPassword}
+                  />
+                  <Route path="/dayview/:date?" component={Dayview} />
+                  <Route path="/weekview" component={Weekview} />
+                  <Route path="/monthview" component={Monthview} />
+                  <Route path="/trackview/:date?/:track?" component={Trackview} />
+                  <Route path="/scheduling/:date?" component={Scheduling} />
+                  <Route path="/tablet" component={RangeOfficerView} />
+                  <Route path="/profile" component={Profile} />
+                  <Route path="/usermanagement" component={UserManagementView} />
+                  <Route path="/supervisor-raffle" component={Raffle} />
+                  <Route path="/tracks" component={TrackCRUD} />
+                  <Route path="/email-settings" component={EmailSettings} />
+                  <Route path="/statistics" component={Statistics} />
+                </Switch>
+              </header>
+            </div>
+          </ThemeProvider>
         </Router>
       </CookiesProvider>
     );
