+++ front/src/resetPW/ResetPassword.js
@@ -2,10 +2,6 @@ import React, { useState } from 'react';
 import classNames from 'classnames';
 
 // Material UI components
-import {
-  createMuiTheme,
-  ThemeProvider,
-} from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
@@ -18,19 +14,6 @@ import translations from '../texts/texts.json';
 import css from './ResetPassword.module.scss';
 
 const classes = classNames.bind(css);
-
-/*
-  TO DO: This same theme is used in SignIn.js
-  -Find out if this theme can be applied to both without two separate declarations
-  (without causing side-effects to other parts of the app)
-*/
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      main: '#555555',
-    },
-  },
-});
 
 const ResetPassword = () => {
   const lang = localStorage.getItem('language');
@@ -68,32 +51,30 @@ const ResetPassword = () => {
   };
 
   return (
-    <ThemeProvider theme={theme}>
-      <Container component="main" maxWidth="xs">
-        <CssBaseline />
-        <div className={classes(css.paper)}>
-          {showForm && (
-            <ResetPasswordForm
-              showNullError={showNullError}
-              showError={showError}
-              isWaiting={isWaiting}
-              onSubmit={reset}
-            />
-          )}
+    <Container component="main" maxWidth="xs">
+      <CssBaseline />
+      <div className={classes(css.paper)}>
+        {showForm && (
+          <ResetPasswordForm
+            showNullError={showNullError}
+            showError={showError}
+            isWaiting={isWaiting}
+            onSubmit={reset}
+          />
+        )}
 
-          {emailSent && (
-            <Typography
-              component="h3"
-              variant="h5"
-              align="center"
-              className={classes(css.success)}
-            >
-              {resetPW.Sent[lang]}
-            </Typography>
-          )}
-        </div>
-      </Container>
-    </ThemeProvider>
+        {emailSent && (
+          <Typography
+            component="h3"
+            variant="h5"
+            align="center"
+            className={classes(css.success)}
+          >
+            {resetPW.Sent[lang]}
+          </Typography>
+        )}
+      </div>
+    </Container>
   );
 };
